backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "openai>=1.100.2",
    "python-dotenv>=1.1.1",
    "youtube-transcript-api>=1.2.2",
]
